further work on dataset refactor
