+++ lib/data/toysets.py
@@ -36,6 +36,59 @@ SD_VARIANT_DATASETS = ["G2"]
 NUM_VARIANT_DATASETS = ["S_set", "A_set"]
 
 
+def HANDLE():
+    if CONFIG.toy_data_path is None:
+        raise ValueError('torchvision dataset must have corresponding torchvision folder specified in '
+                         '`config.yaml`')
+    Dataset = getattr(toysets, source)
+
+    if copy_to_local:
+        copy_to_local_path(path.join(CONFIG.toy_data_path, source))
+        base_path = CONFIG.local_data_path
+    else:
+        base_path = CONFIG.toy_data_path
+
+    train_path = path.join(base_path, source)
+    test_path = train_path
+
+    Dataset = make_indexing(Dataset)
+    Dataset = make_tds_random_and_split(Dataset)
+    train_set = Dataset(root=train_path, download=True, split=0.8, load=True)
+    test_set = Dataset(root=test_path, download=True, split=1, idx=train_set.idx, load=True)
+    output_sources = ['images', 'targets']
+
+    dim_x = train_set.tensors[0].size()[1]
+    dim_l = len(np.unique(np.concatenate([train_set.tensors[1], test_set.tensors[1]])))
+    dims = dict(x=dim_x, labels=dim_l)
+
+
+def make_tds_random_and_split(C):
+    '''Wraps Toyset class to add random splitting.
+
+    Args:
+        C: Toyset data class to be wrapped
+
+    Returns:
+        RandomSplitting class that wraps Toyset data class
+
+    '''
+    class RandomSplitting(C):
+        def __init__(self, *args, idx=None, split=.8, **kwargs):
+            super().__init__(*args, **kwargs)
+            self.idx = idx if idx is not None else torch.randperm(len(self))
+            tensors_ = []
+
+            for i in range(len(self.tensors)):
+                if split > 0:
+                    tensors_.append(self.tensors[i][self.idx][:int(split * len(self))])
+                else:
+                    tensors_.append(self.tensors[i][self.idx][int(split * len(self)) - 1:])
+
+            self.tensors = tuple(tensors_)
+
+    return RandomSplitting
+
+
 class _SmallDataset(data.TensorDataset):
 
     def __init__(self, root, *select,
