+++ lib/entrypoints.py
@@ -1,0 +1,4 @@
+'''Module for entrypoints
+
+'''
+
