+++ lib/plugins.py
@@ -1,0 +1,5 @@
+'''Module for plugins
+
+'''
+
+from .data import DatasetPlugin
